Memory corruption

+++ lib/vecindades.py
@@ -4,7 +4,7 @@ import check
 
 def move_1 (instance, solucion):
     scale,nconf,nchip,max_iter,n_pos,t_ext,tmax_chip,t_delta,tam,chip_info = instance
-    result = np.empty(0)
+    result = np.empty(0,dtype=np.int32)
     
     for i in range(nchip):
         x = solucion[2*i]
